feat: Implement initial project structure including entities, repositories, services, controllers, and migrations for word management, user authentication, and practice features.

+++ WordMaster.Application/Persistence/Repositories/IWordFolderRepository.cs
@@ -3,9 +3,9 @@ namespace WordMaster.Application.Persistence.Repositories;
 
 public interface IWordFolderRepository : IGenericRepository<WordFolder>
 {
-    Task<List<Word>> GetWordsInFolderAsync(int folderId);
-    Task<bool> LinkExistsAsync(int folderId, int wordId);
-    Task<WordFolder?> GetLinkAsync(int folderId, int wordId);
+    Task<List<Word>> GetWordsInFolderAsync(long folderId);
+    Task<bool> LinkExistsAsync(long folderId, long wordId);
+    Task<WordFolder?> GetLinkAsync(long folderId, long wordId);
 }
