+++ WordMaster.Infrastructure/EfCore/Repositories/WordFolderRepository.cs
@@ -6,7 +6,7 @@ namespace WordMaster.Infrastructure.EfCore.Repositories;
 
 public class WordFolderRepository(AppDbContext context) : GenericRepository<WordFolder>(context), IWordFolderRepository
 {
-    public async Task<List<Word>> GetWordsInFolderAsync(int folderId)
+    public async Task<List<Word>> GetWordsInFolderAsync(long folderId)
     {
         return await _context.WordFolders
             .AsNoTracking()
@@ -16,12 +16,12 @@ public class WordFolderRepository(AppDbContext context) : GenericRepository<Word
             .ToListAsync();
     }
 
-    public async Task<bool> LinkExistsAsync(int folderId, int wordId)
+    public async Task<bool> LinkExistsAsync(long folderId, long wordId)
     {
         return await _context.WordFolders.AnyAsync(x => x.FolderId == folderId && x.WordId == wordId);
     }
 
-    public async Task<WordFolder?> GetLinkAsync(int folderId, int wordId)
+    public async Task<WordFolder?> GetLinkAsync(long folderId, long wordId)
     {
         return await _context.WordFolders.FirstOrDefaultAsync(x => x.FolderId == folderId && x.WordId == wordId);
     }
